requirements.txt
click==6.6
coreapi==2.0.0
coreapi-cli==1.0.0
Django==1.10.1
djangorestframework==3.4.6
httpie==0.9.6
itypes==1.1.0
Pygments==2.1.3
requests==2.11.1
uritemplate==3.0.0
datetime==4.1.1